have QueryBuilders extend a BaseBuilder
