+++ src/Query/Builder.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Silk\Query;
+
+use Silk\Type\Model;
+use Silk\Contracts\BuildsQueries;
+
+abstract class Builder implements BuildsQueries
+{
+    /**
+     * Set the model for this query.
+     *
+     * @param Model $model
+     *
+     * @return $this
+     */
+    public function setModel(Model $model)
+    {
+        $this->model = $model;
+
+        return $this;
+    }
+
+    /**
+     * Get the model
+     *
+     * @return Model
+     */
+    public function getModel()
+    {
+        return $this->model;
+    }
+}
